requirements.txt
asgiref==3.10.0
Django==5.2.7
django-filter==25.2
django-jazzmin==3.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.9
pillow==12.0.0
PyJWT==2.10.1
sqlparse==0.5.3
tzdata==2025.2
